% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{add_col}
\alias{add_col}
\title{Add an element in the current cell, and advance to next column}
\usage{
add_col(table_builder, label = NA, subrow = NA, subcol = NA, ...)
}
\description{
Add an element in the current cell, and advance to next column
}

