% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-post.R
\name{add_footnote}
\alias{add_footnote}
\title{Add a footnote to a table}
\usage{
add_footnote(table, footnote)
}
\arguments{
\item{table}{tangram; the tangram table to modify}

\item{footnote}{character; The footnote to add}
}
\value{
the modified table
}
\description{
Add a footnote to a table
}
