% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{add_row}
\alias{add_row}
\title{Add an element in the current cell, and advance down to the next row}
\usage{
add_row(table_builder, label = NA, subrow = NA, subcol = NA, ...)
}
\description{
Add an element in the current cell, and advance down to the next row
}

