% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{args_flatten}
\alias{args_flatten}
\title{Special flatten list, that doesn't flatten cells.
Helper for dealing with ... arguments.}
\usage{
args_flatten(ls)
}
\description{
Special flatten list, that doesn't flatten cells.
Helper for dealing with ... arguments.
}

