% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-typing.R
\name{as.categorical}
\alias{as.categorical}
\title{Convert data type to a factor if it's not already}
\usage{
as.categorical(x)
}
\arguments{
\item{x}{Data to convert to factor}
}
\value{
Data as a factor
}
\description{
Convert data type to a factor if it's not already
}
\examples{

as.categorical(1:3)

}
