% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\docType{class}
\name{ASTBranch}
\alias{ASTBranch}
\title{A left/right branch in an Abstract Syntrax Tree. This inherits from ASTNode, and
is intended to be a base class as well. Should never be instantiated directly
as once again the semantic information is contained in the class name.}
\format{\code{\link{R6Class}} object.}
\usage{
ASTBranch
}
\description{
A left/right branch in an Abstract Syntrax Tree. This inherits from ASTNode, and
is intended to be a base class as well. Should never be instantiated directly
as once again the semantic information is contained in the class name.
}
\section{Fields}{

\describe{
\item{\code{left}}{A pointer to the left node below this one}

\item{\code{right}}{A pointer to the right node below this one}
}}

\section{Methods}{

\describe{
  \item{\code{distribute()}}{Depth first application of distribute() to left and right nodes, then modifies left and right and returns self.}
  \item{\code{terms()}}{Returns the node}
  \item{\code{string()}}{Returns the string formula of the node}
  \item{\code{reduce(data)}}{Given a set of data, perform the logical reduction of the current node.}
}
}

\keyword{data}
