% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\docType{class}
\name{ASTNode}
\alias{ASTNode}
\title{A Node in an Abstract Syntax Tree (AST)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ASTNode
}
\description{
This is the root R6 class of any term of the AST which is created
when parsing a table formula. This should only be used as a base class
as the class information carries the semantic meaning of a given node.
}
\section{Fields}{

\describe{
\item{\code{symbol}}{A string which tells what this node in the AST represents.}

\item{\code{value}}{A string of addtional information contained by the node.}
}}

\section{Methods}{

\describe{
  \item{\code{terms()}}{Returns the node itself}
  \item{\code{distribute()}}{Applies the distributive property to the node, and returns the resulting node.}
  \item{\code{string()}}{Returns the string formula of the node}
  \item{\code{reduce(data)}}{Given a set of data, perform the logical reduction of the current node.}
}
}

\keyword{data}
