% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\docType{class}
\name{ASTPlus}
\alias{ASTPlus}
\title{The addition of two terms, in an ASTNode.}
\format{
\code{\link{R6Class}} object.
}
\description{
The addition of two terms, in an ASTNode.

The addition of two terms, in an ASTNode.
}
\examples{
ASTPlus$new(ASTVariable$new("x"), ASTVariable$new("y"))$string()

}
\keyword{data}
\section{Super classes}{
\code{\link[tangram:ASTNode]{tangram::ASTNode}} -> \code{\link[tangram:ASTBranch]{tangram::ASTBranch}} -> \code{ASTPlus}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Just returns the R6 name 'ASTPlus'}

\item{\code{left}}{The node to the left of this node}

\item{\code{right}}{The node to the right of this node}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ASTPlus$new()}}
\item \href{#method-terms}{\code{ASTPlus$terms()}}
\item \href{#method-string}{\code{ASTPlus$string()}}
\item \href{#method-clone}{\code{ASTPlus$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tangram" data-topic="ASTNode" data-id="set_format">}\href{../../tangram/html/ASTNode.html#method-set_format}{\code{tangram::ASTNode$set_format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tangram" data-topic="ASTBranch" data-id="distribute">}\href{../../tangram/html/ASTBranch.html#method-distribute}{\code{tangram::ASTBranch$distribute()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tangram" data-topic="ASTBranch" data-id="reduce">}\href{../../tangram/html/ASTBranch.html#method-reduce}{\code{tangram::ASTBranch$reduce()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct a new node that represents addition
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ASTPlus$new(left, right)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{left}}{(ASTNode) Node on the left side of the addition}

\item{\code{right}}{(ASTNode) Node on the right side of the addition}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-terms"></a>}}
\if{latex}{\out{\hypertarget{method-terms}{}}}
\subsection{Method \code{terms()}}{
Returns a vector of the left and right terms
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ASTPlus$terms()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-string"></a>}}
\if{latex}{\out{\hypertarget{method-string}{}}}
\subsection{Method \code{string()}}{
A reparsable string representation of this node.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ASTPlus$string()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ASTPlus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
