% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{carriage_return}
\alias{carriage_return}
\title{Goto first column, does not advance to next row}
\usage{
carriage_return(table_builder)
}
\description{
Goto first column, does not advance to next row
}

