% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-operators.R
\name{cbind.tangram}
\alias{cbind.tangram}
\title{A cbind for generated table tangram objects.}
\usage{
\method{cbind}{tangram}(..., deparse.level = 1)
}
\arguments{
\item{...}{tangram objects to cbind}

\item{deparse.level}{numeric; not used}
}
\value{
A merged tangram object
}
\description{
Execute the equivalent of an cbind for generated tables
}
