% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-cell.R
\name{cell.htest}
\alias{cell.htest}
\title{Construct hypothesis test cell}
\usage{
\method{cell}{htest}(x, format = 2, pformat = "\%1.3f", ...)
}
\arguments{
\item{x}{The htest object to convert to a rendereable cell}

\item{...}{additional specifiers for identifying this cell (see key)}
}
\value{
an S3 rendereable cell that is a hypothesis test
}
\description{
Construct a cell from a hypothesis test
}
\examples{
tg(t.test(rnorm(10),rnorm(10)), list(value="A"), list(value="B"))
}
