% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-cell.R
\name{cell_chi2}
\alias{cell_chi2}
\title{Create an cell_chi2 (S3) object of the given statistic}
\usage{
cell_chi2(chi2, df, p, class = NULL, ...)
}
\arguments{
\item{chi2}{The value of the X^2 statistic}

\item{df}{degrees of freedom}

\item{p}{p-value of resulting test}

\item{class}{character; An optional field for additional S3 classes (e.g. could be used in html rendering for CSS)}

\item{...}{optional extra information to attach}
}
\value{
A cell_chi2 object.
}
\description{
A cell_chi2 object contains a statistical result of an X^2-test.
}
\examples{
cell_chi2(5.6, 2, 0.06081)
}
