% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-Cell.R
\name{cell_fraction}
\alias{cell_fraction}
\title{Create an cell_fraction (S3) object of the given statistic}
\usage{
cell_fraction(numerator, denominator, ratio, percentage, src = NA)
}
\arguments{
\item{numerator}{The value of the numerator}

\item{denominator}{The value of the denominator}

\item{ratio}{The ratio of the two}

\item{percentage}{The percentage this represents}

\item{src}{An optional field for traceability of the source of the field}
}
\value{
A cell_fraction object.
}
\description{
A cell_fraction object contains a statistical result of a fraction/percentage.
}
\examples{
cell_fraction(1, 4, 0.25, 25)
}
