% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-Cell.R
\name{cell_fstat}
\alias{cell_fstat}
\title{Create an cell_fstat (S3) object of the given statistic}
\usage{
cell_fstat(f, n1, n2, p, reference = NA, src = NA)
}
\arguments{
\item{f}{The value of the f-statistic}

\item{n1}{1st dimension degrees of freedom}

\item{n2}{2nd dimension degrees of freedom}

\item{p}{The p-value of the resulting test}

\item{reference}{A possible reference number for use in a table key}

\item{src}{An optional field for traceability of the source of the field}
}
\value{
A cell_fstat object.
}
\description{
A cell_fstat object contains a statistical result of an F-test.
}
\examples{
cell_fstat(4.0, 10, 20, 0.004039541, 1, "example")
}
