% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-cell.R
\name{cell_label}
\alias{cell_label}
\title{Create an cell_label (S3) object of the given text.}
\usage{
cell_label(text, units = NULL, class = NULL, ...)
}
\arguments{
\item{text}{character; The text of the label. May include a subset of LaTeX greek or math.}

\item{units}{character; An optional field that contains units}

\item{class}{character; An optional field for additional S3 classes (e.g. could be used in html rendering for CSS)}

\item{...}{optional extra information to attach}
}
\value{
A tangram object
}
\description{
A cell_label object represents a label cell inside a table. It can
also contain units.
}
\examples{
cell_label("Compaction Method")
cell_label("Concentration", "mg/dl")
cell_label("Concentration", "mg/dl", subcol="A")
}
