% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-Cell.R
\name{cell_n}
\alias{cell_n}
\title{Create an cell_n (S3) object of the given statistic}
\usage{
cell_n(n, src = NA)
}
\arguments{
\item{n}{The value of the X^2 statistic}

\item{src}{An optional field for traceability of the source of the field}
}
\value{
A cell_n object.
}
\description{
A cell_n object contains a statistical result of an X^2-test.
}
\examples{
cell_n(20)
}
