% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-cell.R
\name{cell_named_values}
\alias{cell_named_values}
\title{Create named value cells}
\usage{
cell_named_values(values, names, aspect = NULL, sep = ", ", ...)
}
\arguments{
\item{values}{vector; to create cell values from}

\item{names}{character; names to apply to values}

\item{sep}{character; separator to use when rendering}

\item{...}{additional attributes to attach to cell}
}
\value{
A cell object with named values
}
\description{
A cell object with
additionally contain an interval with a low and high of a
specified width.
}
\examples{
cell_named_values(1.0, "one")
}
