% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-Cell.R
\name{cell_quantile}
\alias{cell_quantile}
\title{Create a cell_quantile (S3) object of the given quantile}
\usage{
cell_quantile(quantiles, src = NA)
}
\arguments{
\item{quantiles}{The raw quantiles object.}

\item{src}{An optional field for traceability of the source of the field}
}
\value{
A cell_quantile object.
}
\description{
A cell_quantile object contains the raw quantiles expected
for later rendering.
}
\examples{
require(stats)
cell_quantile(quantile(rnorm(100)), 'Example')
}
