% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-cell.R
\name{cell_range}
\alias{cell_range}
\title{Create a cell representing a range}
\usage{
cell_range(low, high, class = NULL, sep = ", ", ...)
}
\arguments{
\item{low}{character or numeric; lower value of range}

\item{high}{character or numeric; upper value of range}

\item{class}{character; An optional field for additional S3 classes (e.g. could be used in html rendering for CSS)}

\item{sep}{character; separator to use when rendering}

\item{...}{additional attributes to attach to cell}
}
\value{
A cell object denoting a range
}
\description{
Useful for things such as confidence intervals.
}
\examples{
cell_range(-1.0, 1.0)
}
