% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-Cell.R
\name{cell_spearman}
\alias{cell_spearman}
\title{Create an cell_spearman (S3) object of the given statistic}
\usage{
cell_spearman(S, rho, p, reference = NA, src = NA)
}
\arguments{
\item{S}{The value of the spearman statistic}

\item{rho}{The rho value of the test}

\item{p}{p-value of resulting test}

\item{reference}{A possible reference number for use in a table key}

\item{src}{An optional field for traceability of the source of the field}
}
\value{
A cell_spearman object.
}
\description{
A cell_spearman object contains a statistical result of an spearman-test.
}
\examples{
cell_spearman(20, 0.2, 0.05)
}
