% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-cell.R
\name{cell_studentt}
\alias{cell_studentt}
\title{Create an cell_studentt (S3) object of the given statistic}
\usage{
cell_studentt(t, df, p, aspect = NULL, ...)
}
\arguments{
\item{t}{The value of the X^2 statistic}

\item{df}{degrees of freedom}

\item{p}{p-value of resulting test}
}
\value{
A cell_studentt object.
}
\description{
A cell_studentt object contains a statistical result of an t-test.
}
\examples{
cell_studentt(2.0, 20, 0.02963277)
}
