% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-Cell.R
\name{cell_subheader}
\alias{cell_subheader}
\title{Create a cell_subheader (S3) object of the given text.}
\usage{
cell_subheader(text, units = NA, src = NA)
}
\arguments{
\item{text}{The text of the label}

\item{units}{An optional field that contains units}

\item{src}{An optional field for traceability of the source of the field}
}
\value{
A cell_subheader object.
}
\description{
A cell_subheader object is a sub class of a cell_header representing
that this label is a header and upon rendering may invoke different
handling.
}
\examples{
cell_subheader("Italic this")
}
