% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-Cell.R
\name{cell_table}
\alias{cell_table}
\title{Create an empty cell_table}
\usage{
cell_table(rows, cols, embedded = TRUE)
}
\arguments{
\item{rows}{An integer of the number of rows to create}

\item{cols}{An integer of the number of cols to create}

\item{embedded}{A boolean representing whether this table will is to be marked as embedded in another}
}
\value{
An empty cell_table object.
}
\description{
Create an empty cell_table (S3) object to fill with desired table elements. Note that
The initial size is not a limiting factor, the table can grow as needed later.
}
