% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cell_transform}
\alias{cell_transform}
\title{Apply a function to all cells of a table}
\usage{
cell_transform(FUN, ...)
}
\arguments{
\item{table}{the table to modify}

\item{function}{to apply, must return the modified cell}
}
\value{
the modified table
}
\description{
Apply a function to all cells of a table
}

