% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{col_header}
\alias{col_header}
\title{Create a new column header in a table}
\usage{
col_header(table_builder, ..., sub = TRUE)
}
\arguments{
\item{table_builder}{The table builder object to modify}

\item{...}{All the column header elements to add}

\item{sub}{treat as subheader if after first header, defaults to TRUE}
}
\value{
the modified table_builder
}
\description{
Function to append a column header to a table being built. The first call creates
a column header, subsequent calls add sub headers to existing column header
}
