% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-csv.R
\name{csv}
\alias{csv}
\alias{csv.tangram}
\alias{csv.default}
\title{Generate an csv from a tangram or cell object}
\usage{
csv(object, ...)

\method{csv}{tangram}(object, file = NULL, sep = ",", ...)

\method{csv}{default}(object, ...)
}
\arguments{
\item{object}{The cell header to render to HTML5}

\item{...}{additional arguments to renderer. Unused}

\item{file}{File to write result into}

\item{sep}{separator to use}
}
\value{
A string containing the csv file
}
\description{
Given a tangram object create an index representation.
}
