% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{cursor_down}
\alias{cursor_down}
\title{Move cursor down}
\usage{
cursor_down(table_builder, n = 1)
}
\arguments{
\item{table_builder}{The table builder to work on}

\item{n}{units to move cursor down}
}
\value{
a table builder with the cursor down n positions
}
\description{
Move table builder cursor down specified value (default 1)
}
\examples{
library(magrittr)
x <- Parser$new()$run(y ~ x)
new_table_builder(x$right, x$left) \%>\% cursor_pos(3,3) \%>\% cursor_down(2)

}
