% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{cursor_left}
\alias{cursor_left}
\title{Move table builder cursor left the specified value (default 1)}
\usage{
cursor_left(table_builder, n = 1)
}
\arguments{
\item{table_builder}{The table builder to work on}

\item{n}{units to move cursor left}
}
\value{
a table builder with the cursor left n positions
}
\description{
Move table builder cursor left the specified value (default 1)
}
\examples{
x <- Parser$new()$run(y ~ x)
new_table_builder(x$right, x$left) \%>\% cursor_pos(3,3) \%>\% cursor_left(2)

}

