% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{cursor_pos}
\alias{cursor_pos}
\title{Move cursor to position}
\usage{
cursor_pos(table_builder, nrow, ncol)
}
\arguments{
\item{table_builder}{The table builder to work on}

\item{nrow}{The number of the row to move too}

\item{ncol}{The number of the col to move too}
}
\value{
a table builder with the cursor at the specified position
}
\description{
Move table builder cursor to the specified position
}
\examples{
library(magrittr)
x <- Parser$new()$run(y ~ x)
new_table_builder(x$right, x$left) \%>\% cursor_pos(3,3)

}
