% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{cursor_right}
\alias{cursor_right}
\title{Move cursor right}
\usage{
cursor_right(table_builder, n = 1)
}
\arguments{
\item{table_builder}{The table builder to work on}

\item{n}{units to move cursor right}
}
\value{
a table builder with the cursor right n positions
}
\description{
Move table builder cursor right the specified value (default 1)
}
\examples{
library(magrittr)
x <- Parser$new()$run(y ~ x)
new_table_builder(x$right, x$left) \%>\% cursor_pos(3,3) \%>\% cursor_right(2)

}
