% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html5.R
\name{custom_css}
\alias{custom_css}
\title{Return a CSS file as a string}
\usage{
custom_css(filename, id = NA)
}
\arguments{
\item{filename}{Name of the CSS file to load}

\item{id}{CSS id to prepend to all entries}
}
\value{
String of possibly modified CSS file
}
\description{
Given a filename, this function will load the file name from the current working directory.
If it is not found from the current working directory it will search in the package for a
a matching filename and load that instead. If an id is specified, that will be prepended
to all CSS selectors (TODO: make this substitution more robust). The result is returned
as a string.
}
\examples{

custom_css("lancet.css", "tbl1")

}
