% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-post.R
\name{del_col}
\alias{del_col}
\title{Delete a given column from a table}
\usage{
del_col(table, col)
}
\arguments{
\item{table}{the table to modify}

\item{col}{the number of the column to drop}
}
\value{
the modified table
}
\description{
Given a table, remove the specified column
}
