% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-post.R
\name{del_row}
\alias{del_row}
\title{Delete a given row from a table}
\usage{
del_row(table, row)
}
\arguments{
\item{table}{the table to modify}

\item{row}{the number of the row to drop}
}
\value{
the modified table
}
\description{
Given a table, remove the specified row
}
