% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmisc.R
\name{derive_label}
\alias{derive_label}
\title{Determine the label of a given AST node. Should have data attached via reduce before calling.}
\usage{
derive_label(node)
}
\arguments{
\item{node}{Abstract syntax tree node.}
}
\value{
A string with a label for the node
}
\description{
Determine the label of a given AST node. Should have data attached via reduce before calling.
}
\examples{

hmisc_data_type(c(1,2,3))
hmisc_data_type(factor(c("A","B","C")))
hmisc_data_type(factor(c("A","B","B","A")))
hmisc_data_type(factor(c(TRUE, FALSE, TRUE, FALSE)))

}

