% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-post.R
\name{drop_statistics}
\alias{drop_statistics}
\title{Drop all statistics columns from a table.}
\usage{
drop_statistics(table)
}
\arguments{
\item{table}{the table to remove statistical columns}
}
\value{
the modified table
}
\description{
Delete from a table all columns that contain statistics
}
