% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{form}
\alias{form}
\title{Attach format attribute}
\usage{
form(x, value)
}
\arguments{
\item{x}{the object to attach format information}

\item{format}{the formatting to be applied (usually comes from AST node)}
}
\value{
an S3 object with format attr set
}
\description{
Attach formatting information to an object in the attr "format"
}
\examples{
format(2, "\%.2f")
}

