% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-format.R
\name{format_guess}
\alias{format_guess}
\title{Guess the best format for a given set of numerical data}
\usage{
format_guess(x)
}
\arguments{
\item{x}{numeric; basic math and quantile function must work on data passed in}
}
\value{
numeric; the digits past the decimal recommended for display
}
\description{
Given a vector of data, default to 3 significant digits or all if maximum is greater
than zero
}
\examples{
format_guess(rnorm(100))
format_guess(rnorm(100, sd=1e-6))
}
