% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-hmisc.R
\name{hmisc_fraction}
\alias{hmisc_fraction}
\title{Create an cell_fraction (S3) object of the given statistic}
\usage{
hmisc_fraction(numerator, denominator, format = 3, ...)
}
\arguments{
\item{numerator}{numeric; The value of the numerator}

\item{denominator}{numeric; The value of the denominator}

\item{format}{numeric or character; a string formatting directive}

\item{...}{optional extra information to attach}

\item{class}{character; An optional field for additional S3 classes (e.g. could be used in html rendering for CSS)}
}
\value{
A cell_fraction object.
}
\description{
A hmisc_fraction object contains a statistical result of a fraction/percentage.
}
\examples{
hmisc_fraction(1, 4, 3)
}
