% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-post.R
\name{hmisc_intercept_cleanup}
\alias{hmisc_intercept_cleanup}
\title{Cleanup an intercept only model}
\usage{
hmisc_intercept_cleanup(table)
}
\arguments{
\item{table}{the table to modify}
}
\value{
the modified table
}
\description{
Cleanup an intercept only table that was generated from the hmisc default
transform. This drops the statistics column, and modifies the header
to eliminate blank space.
}
