% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-hmisc.R
\name{hmisc_p}
\alias{hmisc_p}
\title{A function to format p values intelligently}
\usage{
hmisc_p(p, pformat = "\%1.3f", include_p = TRUE)
}
\arguments{
\item{p}{numeric; p-value to format}

\item{pformat}{numeric or character; Significant digits or fmt to pass to sprintf}

\item{include_p}{include the leading P on the output string}
}
\value{
A formatted string of the p-value
}
\description{
Given a style in number of digits or a sprintf style specifier it renders
the p-value and checks to see if it's all zeros, then switches the
output to a less than.
}
\examples{
hmisc_p(1e-6)
hmisc_p(0.234)
hmisc_p(1.234e-6, 5)
hmisc_p(1.234e-6, 6)
}
