% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmisc.R
\docType{data}
\name{hmisc_style}
\alias{hmisc_style}
\title{Style Bundle for Hmisc defaults}
\format{An object of class \code{list} of length 4.}
\usage{
hmisc_style
}
\description{
Style Bundle for Hmisc defaults
}
\keyword{data}

