% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-hmisc.R
\name{hmisc_wilcox}
\alias{hmisc_wilcox}
\title{Create an hmisc_wilcox (S3) object of the given statistic}
\usage{
hmisc_wilcox(V, p, class = NULL, ...)
}
\arguments{
\item{p}{p-value of resulting test}

\item{class}{character; An optional field for additional S3 classes (e.g. could be used in html rendering for CSS)}

\item{...}{optional extra information to attach}

\item{S}{The value of the spearman statistic}
}
\value{
A cell_spearman object.
}
\description{
A hmisc_wilcox object contains a statistical result of an Wilcoxon-test.
}
\examples{
hmisc_wilcox(20, 0.2)
}
