% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{home}
\alias{home}
\title{Return table builder cursor position to (1,1)}
\usage{
home(table_builder)
}
\arguments{
\item{table_builder}{The table builder to work on}
}
\value{
a table builder with the cursor at home
}
\description{
Return table builder cursor position to (1,1)
}
\examples{
x <- Parser$new()$run(y ~ x)
new_table_builder(x$right, x$left) \%>\% home()

}

