% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-html5.R
\name{html5.cell}
\alias{html5.cell}
\title{Convert an abstract cell object into an HTML5 string}
\usage{
\method{html5}{cell}(object, id, ..., class = NULL)
}
\arguments{
\item{object}{The cell to render to HTML5}

\item{id}{A unique identifier for traceability}

\item{...}{additional arguments to renderer. Unused}

\item{class}{An additional class attribute for the HTML5 element}
}
\value{
A text string rendering of the given cell as a <td> with several <span>'s.
}
\description{
Given a cell class create an HTML5 representation.
}
