% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html5.R
\name{html5.cell_chi2}
\alias{html5.cell_chi2}
\title{Convert an abstract cell_chi2 object into an HTML5 string}
\usage{
\method{html5}{cell_chi2}(object, caption, ..., class = NA)
}
\arguments{
\item{object}{The cell chi2 to render to HTML5}

\item{caption}{A string caption for the table}

\item{...}{additional arguments to renderer. Unused}

\item{class}{An additional class attribute for the HTML5 element}
}
\value{
A text string rendering of the given chi2 as a <td> with several <span>'s.
}
\description{
Given a cell_chi2 class create an HTML5 representation.
}
