% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html5.R
\name{html5.cell_estimate}
\alias{html5.cell_estimate}
\title{Convert an abstract cell_estimate object into an HTML5 string}
\usage{
\method{html5}{cell_estimate}(object, caption, ..., class = NA)
}
\arguments{
\item{object}{The cell estimate to render to HTML5}

\item{caption}{A string caption for the table}

\item{...}{additional arguments to renderer. Unused}

\item{class}{An additional class attribute for the HTML5 element}
}
\value{
A text string rendering of the given estimate as a <td> with several <span>'s.
}
\description{
Given a cell_estimate class create an HTML5 representation.
}
