% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-html5.R
\name{html5.cell_fraction}
\alias{html5.cell_fraction}
\title{Convert an abstract cell_fraction object into an HTML5 string}
\usage{
\method{html5}{cell_fraction}(object, id, ..., class = NULL)
}
\arguments{
\item{object}{The cell fraction to render to HTML5}

\item{id}{A unique identifier for traceability}

\item{...}{additional arguments to renderer. Unused}

\item{class}{An additional class attribute for the HTML5 element}
}
\value{
A text string rendering of the given fraction as a <td> with several <span>'s.
}
\description{
Given a cell_fraction class create an HTML5 representation.
}
