% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html5.R
\name{html5.cell_subheader}
\alias{html5.cell_subheader}
\title{Convert an abstract cell_subheader object into an HTML5 string}
\usage{
\method{html5}{cell_subheader}(object, caption, ...)
}
\arguments{
\item{object}{The cell subheader to render to HTML5}

\item{caption}{A string caption for the table}

\item{...}{additional arguments to renderer. Unused}
}
\value{
A text string rendering of the given subheader as a <td> with several <span>'s.
}
\description{
Given a cell_subheader class create an HTML5 representation.
}
