% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html5.R
\name{html5.default}
\alias{html5.default}
\title{Default conversion to HTML5 for an abstract table element}
\usage{
\method{html5}{default}(object, caption, ..., class = NA)
}
\arguments{
\item{object}{The cell to render to HTML5}

\item{caption}{A string caption for the table}

\item{...}{additional arguments to renderer. Unused}

\item{class}{An additional class attribute for the HTML5 element}
}
\value{
A text string rendering of the given cell as a <td> with several <span>'s.
}
\description{
Gives a warning and produces an empty cell
}
