% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index.cell_chi2}
\alias{index.cell_chi2}
\title{Generate an an index from a cell_chi2 object}
\usage{
\method{index}{cell_chi2}(object, caption, ...)
}
\arguments{
\item{object}{The cell_chi2 for indexing}

\item{caption}{an additional specifier for the object key}

\item{...}{additional arguments to renderer. Unused}
}
\value{
A matrix of strings containing key, source and value
}
\description{
Given a cell_chi2 class create an index representation. If no source
is specified no index will be generated.
}
