% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index.cell_estimate}
\alias{index.cell_estimate}
\title{Generate an index from a cell_estimate object}
\usage{
\method{index}{cell_estimate}(object, caption, ...)
}
\arguments{
\item{object}{The cell_estimate for indexing}

\item{caption}{an additional specifier for the object key}

\item{...}{additional arguments to renderer. Unused}
}
\value{
A matrix of strings containing key, source and value
}
\description{
Given a cell_estimate class create an index representation. If no source
is specified no index will be generated.
}
