% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index.cell_table}
\alias{index.cell_table}
\title{Generate an an index from a cell_table object}
\usage{
\method{index}{cell_table}(object, caption = "Table", ...)
}
\arguments{
\item{object}{The cell_table for indexing}

\item{caption}{an additional specifier for the object key}

\item{...}{additional arguments to renderer. Unused}
}
\value{
A matrix of strings containing key, source and value
}
\description{
Given a cell_table class create an index representation.
}
