% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-index.R
\name{index.default}
\alias{index.default}
\title{Generate an index from a cell object}
\usage{
\method{index}{default}(object, id = "tangram", name = NULL, key.len = 4,
  ...)
}
\arguments{
\item{object}{cell; The cell for indexing}

\item{id}{character; an additional specifier for the object key}

\item{name}{character; optional names of elements inside object}

\item{key.len}{numeric; length of generated key}

\item{...}{additional arguments to renderer. Unused}
}
\value{
A list of strings containing key, source and value
}
\description{
Given a cell class create an index representation. If no source
is specified no index will be generated.
}
