% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-index.R
\name{index.list}
\alias{index.list}
\title{Generate an index from a list object}
\usage{
\method{index}{list}(object, id = "tangram", key.len = 4, ...)
}
\arguments{
\item{object}{cell; The cell for indexing}

\item{id}{character; an additional specifier for the object key}

\item{key.len}{numeric; length of key to generate}

\item{...}{additional arguments to renderer. Unused}
}
\value{
A list of strings containing key, source and value
}
\description{
Given a cell class create an index representation. If no source
is specified no index will be generated.
}
