% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-post.R
\name{insert_row}
\alias{insert_row}
\title{Insert a row into a tangram table}
\usage{
insert_row(table, after, ..., class = NULL)
}
\arguments{
\item{table}{the table to modify}

\item{after}{numeric; The row to position the new row after. Can be zero for inserting a new first row.}

\item{...}{Table cells to insert. Cannot be larger than existing table.}

\item{class}{character; Classes to apply as directives to renderers}
}
\value{
the modified table
}
\description{
Insert a row into a tangram table. Will fill with empty cells is not enough cells are specified.
}
