% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-typing.R
\name{is.binomial}
\alias{is.binomial}
\title{Determine if a vector is binomial or not}
\usage{
is.binomial(x, threshold = NA)
}
\arguments{
\item{x}{Vector to determine type of}

\item{threshold}{The upper threshold of unique values for which a vector is considered categorical.}
}
\value{
a Boolean: TRUE / FALSE
}
\description{
Determine if a vector is binomial or not
}
\examples{

is.binomial(c(1,2,3))
is.binomial(factor(c("A","B","C")))
is.binomial(factor(c("A","B","B","A")))
is.binomial(factor(c(TRUE, FALSE, TRUE, FALSE)))
is.binomial(c('M', 'F', 'M', 'F'), 10)
}
