% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-lancet.R
\docType{data}
\name{lancet_cell}
\alias{lancet_cell}
\title{Cell Generation functions for Lancet styling}
\format{
An object of class \code{list} of length 8.
}
\usage{
lancet_cell
}
\description{
Each function here is called when a cell is generated. Overriding these in a formula call will allows
one to customize exactly how each cell's contents are generated.
}
\keyword{data}
