% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-latex-map.R
\name{latexify}
\alias{latexify}
\title{LaTeX safe string conversion}
\usage{
latexify(x)
}
\arguments{
\item{x}{string to make LaTeX safe}
}
\value{
valid LaTeX code
}
\description{
LaTeX safe string conversion. This transforms a string handling Markdown characters and UNICODE as best
it can with an automated pass.
}
