% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{line_feed}
\alias{line_feed}
\title{Advance down to next line, does not goto first column}
\usage{
line_feed(table_builder, n = 1)
}
\description{
Advance down to next line, does not goto first column
}

