% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-nejm.R
\name{nejm_fraction}
\alias{nejm_fraction}
\title{Create an cell_fraction (S3) in NEJM style of the given data}
\usage{
nejm_fraction(numerator, denominator, format = NULL, ...)
}
\arguments{
\item{numerator}{numeric; The value of the numerator}

\item{denominator}{numeric; The value of the denominator}

\item{format}{numeric or character; a string formatting directive}

\item{...}{optional extra information to attach}
}
\value{
A cell_fraction object.
}
\description{
A cell object contains a statistical result of a fraction/percentage in nejm style
}
\examples{
nejm_fraction(1, 4, 3)
}
