% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-nejm.R
\name{nejm_range}
\alias{nejm_range}
\title{Create a NEJM style range}
\usage{
nejm_range(x, format, ...)
}
\arguments{
\item{x}{numeric vector whose range is desired}

\item{format}{numeric or character; an argument to pass to the formatting function}

\item{...}{additional arguments to passed to cell()}
}
\description{
Construct a cell which has the range of the given data in NEJM style
}
