% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{new_col}
\alias{new_col}
\title{Advance to the furthest right column on the top row and open a new column}
\usage{
new_col(table_builder)
}
\description{
Advance to the furthest right column on the top row and open a new column
}

