% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{new_header}
\alias{new_header}
\title{Create a new header on a table}
\usage{
new_header(table_builder, attribute, sub, ...)
}
\arguments{
\item{table_builder}{The table builder object to modify}

\item{attribute}{The header attribute name, i.e. row_header or col_header}

\item{sub}{boolean indicating if this is a subheader}

\item{...}{All the header elements to add}
}
\value{
the modified table_builder
}
\description{
Function to append a header object to a given attribute. Will create
a new header if one doesn't exit, or append to existing
}
