% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{new_row}
\alias{new_row}
\title{Advance to the bottom of all defined rows, and open a new one}
\usage{
new_row(table_builder)
}
\description{
Advance to the bottom of all defined rows, and open a new one
}

