% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{new_table_builder}
\alias{new_table_builder}
\title{Create empty table builder.}
\usage{
new_table_builder(row, column)
}
\arguments{
\item{row}{The row node from the AST}

\item{column}{The col node from the AST}
}
\value{
a table builder with 1 empty cell at position (1,1)
}
\description{
Function to create a new table builder to use in continuations.
This maintains a cursor state where values are being written to the
table under construction, as well as references to the row and column
for automated tracability when generating indexes.
}
\examples{
x <- Parser$new()$run(y ~ x)
new_table_builder(x$right, x$left)

}
