% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-hmisc.R
\docType{data}
\name{pbc}
\alias{pbc}
\title{Mayo Clinic Primary Biliary Cirrhosis Data}
\format{An object of class \code{data.frame} with 418 rows and 19 columns.}
\usage{
pbc
}
\description{
D This data is from the Mayo Clinic trial in primary biliary cirrhosis (PBC) of the liver conducted between 1974 and 1984. A total of 424 PBC patients, referred to Mayo Clinic during that ten-year interval, met eligibility criteria for the randomized placebo controlled trial of the drug D-penicillamine. The first 312 cases in the data set participated in the randomized trial and contain largely complete data. The additional 112 cases did not participate in the clinical trial, but consented to have basic measurements recorded and to be followed for survival. Six of those cases were lost to follow-up shortly after diagnosis, so the data here are on an additional 106 cases as well as the 312 randomized participants.
}
\details{
A nearly identical data set found in appendix D of Fleming and Harrington; this version has fewer missing values.

Included for use in example from Hmisc.
}
\keyword{data}
