% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-operators.R
\name{pipe.tangram}
\alias{pipe.tangram}
\alias{|.tangram}
\title{Provide a "|" operator for cbind of tangram tables}
\usage{
\method{|}{tangram}(x, y)
}
\arguments{
\item{x}{left argument for rbind}

\item{y}{right argument for rbind}
}
\value{
A column wise merged tangram object
}
\description{
The pipe operator provides an cbind for tangram tables
}
