% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-operators.R
\name{+.tangram}
\alias{+.tangram}
\title{Provide a "+" operator for rbind of tangram tables}
\usage{
\method{+}{tangram}(x, y)
}
\arguments{
\item{x}{left argument for rbind}

\item{y}{right argument for rbind}
}
\value{
A row wise merged tangram object
}
\description{
The plus operator provides an rbind for tangram tables
}
