% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-summary.R
\name{print.cell}
\alias{print.cell}
\alias{print.tangram}
\title{Print methods for tangram objects}
\usage{
\method{print}{cell}(x, ...)

\method{print}{tangram}(x, ...)
}
\arguments{
\item{x}{object; the item to render}

\item{...}{additional arguments passed to summary}
}
\value{
the text summary
}
\description{
Print methods for tangram objects
}
\examples{
print(cell_label("123"))
print(cell_iqr(rnorm(20)))
print(cell_estimate(2.1,0.8, 3.3))
print(cell_fraction(45, 137))
print(table_builder()   \%>\%
        row_header("row") \%>\%
        col_header(1,2,3) \%>\%
        add_col("A","B","C"))
print(tangram(drug~bili, pbc))
}
