% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-summary.R
\name{print.table_builder}
\alias{print.table_builder}
\title{Print a text summary of a given table_builder}
\usage{
\method{print}{table_builder}(x, ...)
}
\arguments{
\item{x}{The table_builder to render to text}

\item{...}{additional arguments, unused at present}
}
\value{
A text string rendering of the given table
}
\description{
Print a text summary of a given table_builder
}
