% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-proc-tab.R
\name{proc_tab}
\alias{proc_tab}
\title{Tangram transform for proc_tab style summaries via a function}
\usage{
proc_tab(table, row, column, fun = NULL, overall = FALSE, ...)
}
\arguments{
\item{table}{The table builder object}

\item{row}{The row from the abstract syntax tree that parsed the formula}

\item{column}{The column from the abstract syntax tree that parsed the formula}

\item{fun}{The function to apply to the broken out categories}

\item{overall}{Provide a summary of categorical breakdowns}

\item{...}{additional arguments to pass to fun}
}
\description{
Given a function that produces a vector of tangram cells, will generate a table
}
