% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-operators.R
\name{rbind.tangram}
\alias{rbind.tangram}
\title{An rbind for generated tables tangram objects.}
\usage{
\method{rbind}{tangram}(..., deparse.level = 1)
}
\arguments{
\item{...}{tangram objects to rbind}

\item{deparse.level}{numeric; not used}
}
\value{
A merged tangram object
}
\description{
Execute the equivalent of an rbind for generated tables
}
