% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-format.R
\name{render_f}
\alias{render_f}
\title{Format a vector of provided numeric values}
\usage{
render_f(x, format)
}
\arguments{
\item{x}{numeric; the data to format. Must work with quantile function.}

\item{format}{numeric or character; If numeric preserve that many position past the decimal, if character pass directly into sprintf as format string}
}
\value{
character; formatted values as character strings
}
\description{
Given a vector of data return as strings formatted as requested
}
\examples{
render_f(rnorm(5), 3)
render_f(round(rnorm(5), 2), "\%010.03f")
}
