% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-summary.R
\name{render_route_tangram}
\alias{render_route_tangram}
\title{Router for rendering method}
\usage{
render_route_tangram()
}
\value{
A rendering function to use
}
\description{
This functions detects if knitr is loaded, and does it's best to determine the output
format from knitr and returns the appropriate rendering function.
}
