% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-post.R
\name{replace_cell}
\alias{replace_cell}
\title{Replace a cell's contents}
\usage{
replace_cell(table, row, col, object, ...)
}
\arguments{
\item{table}{the tangram table to modify}

\item{row}{numeric; The row to modify}

\item{col}{numeric; The column to modify}

\item{object}{The cell or object to replace in a table}

\item{...}{Additional parameters passed to cell function if not given a cell object}
}
\value{
the modified table
}
\description{
Replace a cell in a table
}
