% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{rmd.cell_table}
\alias{rmd.cell_table}
\title{Generate an Rmd table entry from a cell_table object}
\usage{
\method{rmd}{cell_table}(object, ...)
}
\arguments{
\item{object}{The cell_fstat for indexing}

\item{...}{additional arguments to renderer. Unused}
}
\value{
A string representation of the table
}
\description{
Given a cell_table object generate the corresponding piece of an Rmd table
}
