% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-rmd.R
\name{rmd.tangram}
\alias{rmd.tangram}
\title{Generate an Rmd table entry from a tangram object}
\usage{
\method{rmd}{tangram}(object, ...)
}
\arguments{
\item{object}{The cell_fstat for indexing}

\item{...}{additional arguments to renderer. Unused}
}
\value{
A string representation of the table
}
\description{
Given a tangram object generate the corresponding piece of an Rmd table
}
