% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{row_header}
\alias{row_header}
\title{Create a new row header in a table.}
\usage{
row_header(table_builder, ..., sub = TRUE)
}
\arguments{
\item{table_builder}{The table builder object to modify}

\item{...}{All the row header elements to add}

\item{sub}{treat as subheader if after first, default to TRUE}
}
\value{
the modified table_builder
}
\description{
Function to append a row header to a table being built. The first call creates
a row header, subsequent calls add sub headers to existing row header
}
