% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-rtf.R
\name{rtf.cell}
\alias{rtf.cell}
\title{Given a cell class create an RTF representation.}
\usage{
\method{rtf}{cell}(object, id, ...)
}
\arguments{
\item{object}{The cell to render to RTF}

\item{id}{A unique identifier for traceability}

\item{...}{additional arguments to renderer. Unused}
}
\value{
An RTF string rendering of the given cell.
}
\description{
Given a cell class create an RTF representation.
}
