% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-rtf.R
\name{rtf.cell_chi2}
\alias{rtf.cell_chi2}
\title{Convert an abstract cell_chi2 object into an rtf string}
\usage{
\method{rtf}{cell_chi2}(object, id, ...)
}
\arguments{
\item{object}{The cell chi2 to render to HTML5}

\item{id}{A unique identifier for traceability}

\item{...}{additional arguments to renderer. Unused}
}
\value{
A text string rendering of the given chi2 in rtf
}
\description{
Given a cell_chi2 class create an rtf representation.
}
