% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf.R
\name{rtf.cell_fstat}
\alias{rtf.cell_fstat}
\title{Convert an abstract cell_fstat object into an RTF string}
\usage{
\method{rtf}{cell_fstat}(object, caption, ..., class = NA)
}
\arguments{
\item{object}{The cell fstat to render to RTF}

\item{...}{additional arguments to renderer. Unused}

\item{class}{An additional class attribute for the RTF element}
}
\value{
A text string rendering of the given fstat as a <td> with several <span>'s.
}
\description{
Given a cell_fstat class create an RTF representation.
}
