% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf.R
\name{rtf.cell_header}
\alias{rtf.cell_header}
\title{Convert an abstract cell_header object into an RTF string}
\usage{
\method{rtf}{cell_header}(object, ...)
}
\arguments{
\item{object}{The cell header to render to RTF}

\item{...}{additional arguments to renderer. Unused}
}
\value{
An RTF string rendering of the given header
}
\description{
Given a cell_header class create an RTF representation.
}
