% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-rtf.R
\name{rtf.cell_iqr}
\alias{rtf.cell_iqr}
\title{Convert an abstract cell_iqr object into an RTF string}
\usage{
\method{rtf}{cell_iqr}(object, id, ..., point = 9)
}
\arguments{
\item{object}{The cell quantile to render to RTF}

\item{id}{A unique identifier for traceability}

\item{...}{additional arguments to renderer. Unused}

\item{point}{numeric; The font point size to use in display}
}
\value{
An RTF string rendering of the given quantile.
}
\description{
Given a cell_quantile class create an RTF representation.
}
