% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf.R
\name{rtf.cell_label}
\alias{rtf.cell_label}
\title{Given a cell_label class create an RTF representation.}
\usage{
\method{rtf}{cell_label}(object, ...)
}
\arguments{
\item{object}{The cell label to render to RTF}

\item{...}{additional arguments to renderer. Unused}

\item{caption}{A string caption for the table}
}
\value{
An RTF text string rendering of the given label.
}
\description{
Given a cell_label class create an RTF representation.
}
