% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf.R
\name{rtf.cell_n}
\alias{rtf.cell_n}
\title{Convert an abstract cell_n object into an RTF string}
\usage{
\method{rtf}{cell_n}(object, ...)
}
\arguments{
\item{object}{The cell n to render to RTF}

\item{...}{additional arguments to renderer. Unused}
}
\value{
An RTF string rendering of the given n.
}
\description{
Given a cell_n class create an RTF representation.
}
