% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf.R
\name{rtf.cell_quantile}
\alias{rtf.cell_quantile}
\title{Convert an abstract cell_quantile object into an RTF string}
\usage{
\method{rtf}{cell_quantile}(object, ..., class = NA)
}
\arguments{
\item{object}{The cell quantile to render to RTF}

\item{...}{additional arguments to renderer. Unused}

\item{class}{An additional class attribute for the RTF element}
}
\value{
An RTF string rendering of the given quantile.
}
\description{
Given a cell_quantile class create an RTF representation.
}
