% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-rtf.R
\name{rtf.default}
\alias{rtf.default}
\title{Default conversion to RTF for an abstract table element}
\usage{
\method{rtf}{default}(object, id, ...)
}
\arguments{
\item{object}{The cell to render to RTF}

\item{id}{A unique identifier for traceability}

\item{...}{additional arguments to renderer. Unused}
}
\value{
A RTF string rendering of the given cell
}
\description{
Gives a warning and produces an empty cell
}
