% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-rtf.R
\name{rtf.tangram}
\alias{rtf.tangram}
\title{Convert a tangram into an RTF string or file}
\usage{
\method{rtf}{tangram}(object, id = NA, caption = NA, fragment = FALSE,
  widths = NA, footnote = NA, filename = NA, append = FALSE,
  point = 9, ...)
}
\arguments{
\item{object}{The cell table to render to RTF}

\item{id}{A unique identifier for the table (strongly recommended).}

\item{caption}{A string caption for the table}

\item{fragment}{A boolean flag that determines whether a fragment or a complete RTF document is generatedf}

\item{widths}{RTF requires specified left margin and column widths, this allows user control over these (inches)}

\item{footnote}{Any footnotes to include under the table.}

\item{filename}{A filename to write resulting rtf file to}

\item{append}{A boolean for whether or not to append to given filename}

\item{point}{Main font point size}

\item{...}{additional arguments Fto renderer. Unused}
}
\value{
A text string rendering of the given table
}
\description{
Given a tangram class, a series of conversion creates an rtf
representation of the table.
}
