% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-post.R
\name{select_col}
\alias{select_col}
\title{Select given column(s) from a table}
\usage{
select_col(table, col)
}
\arguments{
\item{table}{the table to modify}

\item{col}{vector containing column(s) to select}
}
\value{
the modified table
}
\description{
Given a table, select the specified column(s)
}
