% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-post.R
\name{select_row}
\alias{select_row}
\title{Select given row(s) from a table}
\usage{
select_row(table, row)
}
\arguments{
\item{table}{the table to modify}

\item{row}{vector with row numbers to select}
}
\value{
the modified table
}
\description{
Given a table, select the specified rows
}
