% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R, R/transform-smd.R
\docType{data}
\name{smd}
\alias{smd}
\title{Compute the standardized mean distance between 2 groups for numerical
or categorical information. Using method described in
'A unifed approach to measuring the effect size between two groups
using SAS' by Dongsheng Yand and Jarrod E. Dalton, 2012. SAS Global
Forum 2012}
\format{An object of class \code{list} of length 5.}
\usage{
smd

smd
}
\arguments{
\item{x}{vector; data to estimate effect size for groups}

\item{group}{vector; the grouping variable.}

\item{weights}{vector; weighting information for x}
}
\description{
Compute the standardized mean distance between 2 groups for numerical
or categorical information. Using method described in
'A unifed approach to measuring the effect size between two groups
using SAS' by Dongsheng Yand and Jarrod E. Dalton, 2012. SAS Global
Forum 2012

List of lists, should contain a "Type" entry with a function to determine type of vector passed in.
Next entries are keyed off returned types from function, and represent the type of a row.
The returned list should contain the same list of types, and represents the type of a column. Thus it now returns
a function to process the intersection of those two types.
}
\keyword{data}
\keyword{datasets}
