% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-smd.R
\docType{data}
\name{smd_cell}
\alias{smd_cell}
\title{Cell Generation functions for SMD comparisons of categorical to numerical}
\format{
An object of class \code{list} of length 7.
}
\usage{
smd_cell
}
\description{
Each function here is called when a cell is generated. Overriding these in a formula call will allows
one to customize exactly how each cell's contents are generated.
}
\details{
While this serves as the base template for transforms, it is by no means required if one develops their
own bundle of data transforms. One can create ay number of cell level styling choices.
}
\keyword{data}
