% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-smd.R
\name{smd_fraction}
\alias{smd_fraction}
\title{Create a fraction cell in the smd transform}
\usage{
smd_fraction(num, den, format, ...)
}
\arguments{
\item{num}{numerator of fraction}

\item{den}{denominator of fraction}

\item{format}{formatting to apply to result}

\item{...}{additional arguments to pass to cell generation}
}
\value{
a tangram cell
}
\description{
Create a fraction cell in the smd transform. In this instance it
print the numerator followed by percentage in parenthesis.
}
