% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-smd.R
\name{smd_meansd}
\alias{smd_meansd}
\title{Create an SMD mean and standard deviation cell}
\usage{
smd_meansd(x, format, ...)
}
\arguments{
\item{x}{vector; variable to evaluate with smd}

\item{format}{formatting to apply to result}

\item{...}{additional arguments to pass to cell generation}
}
\value{
a tangram cell
}
\description{
Create an SMD mean and standard deviation cell. In this case it prints the
mean with the standard deviation in parenthesis
}
