% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{standard_difference}
\alias{standard_difference}
\title{Compute the standardized mean distance between 2 groups for numerical
or categorical information. Using method described in
'A unifed approach to measuring the effect size between two groups
using SAS' by Dongsheng Yand and Jarrod E. Dalton, 2012. SAS Global
Forum 2012}
\usage{
standard_difference(x, group, weight = NULL)
}
\arguments{
\item{x}{vector; data to estimate effect size for groups}

\item{group}{vector; the grouping variable.}

\item{weight}{vector; weighting information for x}
}
\description{
Compute the standardized mean distance between 2 groups for numerical
or categorical information. Using method described in
'A unifed approach to measuring the effect size between two groups
using SAS' by Dongsheng Yand and Jarrod E. Dalton, 2012. SAS Global
Forum 2012
}
