% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-hmisc.R
\name{summarize_chisq}
\alias{summarize_chisq}
\title{Create a summarization for a categorical row versus a categorical column}
\usage{
summarize_chisq(table, row, column, pformat = NULL, collapse_single = TRUE,
  overall = NULL, test = TRUE, row_percents = FALSE, ...)
}
\arguments{
\item{table}{The table object to modify}

\item{row}{The row variable object to use (categorical)}

\item{column}{The column variable to use (categorical)}

\item{pformat}{numeric, character or function; A formatting directive to be applied to p-values}

\item{collapse_single}{logical; default TRUE. Categorical variables with a two values collapse to single row.}

\item{overall}{logical; Include the overall summary column}

\item{test}{logical; include statistical test results}

\item{row_percents}{logical; use denominator across rows instead of columns.}

\item{...}{absorbs extra parameters. Currently unused.}
}
\value{
The modified table object
}
\description{
Given a row and column object from the parser apply a chi^2 test and output
the results
}
