% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-hmisc.R
\name{summarize_kruskal_vert}
\alias{summarize_kruskal_vert}
\title{Create a summarization for a categorical row versus a numerical column}
\usage{
summarize_kruskal_vert(table, row, column, pformat = NULL, ...)
}
\arguments{
\item{table}{The table object to modify}

\item{row}{The row variable object to use (categorical)}

\item{column}{The column variable to use (numerical)}

\item{pformat}{numeric or character; A formatting directive to be applied to p-values}

\item{...}{absorbs additional arugments. Unused at present.}
}
\value{
The modified table object
}
\description{
Given a row and column object from the parser apply a Kruskal test and output
the results vertically (#Categories+1) X (N, Summary, Statistic)
}
