% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmisc.R
\name{summarize_spearman}
\alias{summarize_spearman}
\title{Create a summarization for a numerical row versus a numerical column}
\usage{
summarize_spearman(table, row, column)
}
\arguments{
\item{table}{The table object to modify}

\item{row}{The row variable object to use (numerical)}

\item{column}{The column variable to use (numerical)}
}
\value{
The modified table object
}
\description{
Given a row and column object from the parser apply a Spearman test and output
the results in a 1X3 format.
}
