% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-summary.R
\name{summary.tangram}
\alias{summary.tangram}
\alias{summary.table_builder}
\alias{summary.cell}
\title{The default method for rendering tangram objects
A tangram is a summary, so it returns itself. Otherwise convert to a text representation.}
\usage{
\method{summary}{tangram}(object, ...)

\method{summary}{table_builder}(object, ...)

\method{summary}{cell}(object, ...)
}
\arguments{
\item{object}{object; the item to render}

\item{...}{additional arguments passed to summary}
}
\value{
the text summary
}
\description{
The default method for rendering tangram objects
A tangram is a summary, so it returns itself. Otherwise convert to a text representation.
}
\examples{
summary(cell_label("123"))
summary(hmisc_iqr(rnorm(20)))
summary(hmisc_fraction(45, 137))
summary(table_builder()   \%>\%
        row_header("row") \%>\%
        col_header(1,2,3) \%>\%
        add_col("A","B","C"))
summary(tangram(drug~bili, pbc))
}
