% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-summary.R
\name{summary.table_builder}
\alias{summary.table_builder}
\title{Create a text summary of a given table_builder}
\usage{
\method{summary}{table_builder}(object, ...)
}
\arguments{
\item{object}{The table_builder to render to text}

\item{...}{additional arguments to renderer. Unused at present.}
}
\value{
A text string rendering of the given table
}
\description{
Create a text summary of a given table_builder
}
