% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-summary.R
\name{summary.tangram}
\alias{summary.tangram}
\title{Create a text summary of a given table}
\usage{
\method{summary}{tangram}(object, ...)
}
\arguments{
\item{object}{The cell table to render to text}

\item{...}{additional arguments to renderer. Unused at present.}
}
\value{
A text string rendering of the given table
}
\description{
Create a text summary of a given table
}
