% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{summary_frame}
\alias{summary_frame}
\title{Generate a table using a data frame directly}
\usage{
summary_frame(data, colheader = NA)
}
\arguments{
\item{data}{the data frame to use}

\item{colheader}{vector of headers to use for columns}
}
\value{
table
}
\description{
Transform a data frame directly into a table
}
