% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{summary_table}
\alias{summary_table}
\title{Generate a summary table using a specified formula and data frame}
\usage{
summary_table(formula, data, transforms = hmisc_style, after = NA)
}
\arguments{
\item{formula, }{the formula to apply for summarization}

\item{data}{the data frame to use}

\item{transforms}{a list of lists that contain the transformation to apply for summarization}

\item{after}{one or more functions to further process an abstract table}
}
\value{
the table flattened
}
\description{
Generate a summary table using a specified formula and data frame
}
\examples{

summary_table("drug ~ bili + albumin + stage::Categorical + protime + sex + age + spiders", pbc)

}
