% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-operators.R
\name{col_header}
\alias{col_header}
\alias{row_header}
\alias{write_cell}
\alias{home}
\alias{cursor_up}
\alias{cursor_down}
\alias{cursor_left}
\alias{cursor_right}
\alias{cursor_pos}
\alias{carriage_return}
\alias{line_feed}
\alias{new_line}
\alias{new_row}
\alias{new_col}
\alias{table_apply}
\alias{add_col}
\alias{add_row}
\alias{set_footnote}
\alias{set_id}
\alias{set_caption}
\alias{set_style}
\title{A set of magrittr operators for tangram tables}
\usage{
col_header(table, ..., sub = TRUE)

row_header(table, ..., sub = TRUE)

write_cell(table, x, ...)

home(table)

cursor_up(table, n = 1)

cursor_down(table, n = 1)

cursor_left(table, n = 1)

cursor_right(table, n = 1)

cursor_pos(table, nrow, ncol)

carriage_return(table)

line_feed(table, n = 1)

new_line(table)

new_row(table)

new_col(table)

table_apply(table, X, FUN, ...)

add_col(table, ...)

add_row(table, ...)

set_footnote(table, footnote)

set_id(table, id)

set_caption(table, caption)

set_style(table, style)
}
\description{
A set of magrittr operators for tangram tables
}
