% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{table_builder_apply}
\alias{table_builder_apply}
\title{Nest table continuation inside apply}
\usage{
table_builder_apply(table_builder, X, FUN)
}
\description{
Nest table continuation inside apply
}

