% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{table_flatten}
\alias{table_flatten}
\title{Given a cell_table object with embedded tables, flattens to a single table.}
\usage{
table_flatten(table)
}
\arguments{
\item{table}{the table object to flatten}
}
\value{
the flattened table object
}
\description{
Flattening function to expanded embedded tables inside table cells.
}
