% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-cell.R
\name{tangram}
\alias{tangram}
\title{Create an empty table of cells, a tangram}
\usage{
tangram(rows = 1, cols = 1, embedded = FALSE)
}
\arguments{
\item{rows}{An integer of the number of rows to create}

\item{cols}{An integer of the number of cols to create}

\item{embedded}{logical; Will this table be embedded inside another}
}
\value{
An empty tangram object.
}
\description{
Create an empty list of lists to fill with desired table elements. Note that
The initial size is not a limiting factor, the table can grow as needed later.
}
