% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{tg.cell}
\alias{tg.cell}
\title{Identity function on cell}
\usage{
\method{tg}{cell}(x, row, column, ...)
}
\arguments{
\item{x}{the object to return}

\item{row}{The AST row of that is generating this cell}

\item{column}{The AST column that is generating this cell}

\item{...}{additional specifiers for identifying this cell (see key)}
}
\value{
an S3 rendereable cell label
}
\description{
Construct a cell from a cell. This is essentially an identity function.
I.e., if a user has constructed a cell and passed it in do nothing.
}
\examples{
tg(tg("Joe"), list(value="A"), list(value="B"))
}
