% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{tg.default}
\alias{tg.default}
\title{Default cell is a label}
\usage{
\method{tg}{default}(x, row, column, ...)
}
\arguments{
\item{x}{Object to turn into a renderable label cell}

\item{row}{The AST row of that is generating this cell}

\item{column}{The AST column that is generating this cell}

\item{...}{additional specifiers for identifying this cell (see key)}
}
\value{
an S3 rendereable cell label
}
\description{
Construct a cell. This is the default fallback, it creates a label cell
}
\examples{
tg("Joe", list(value="A"), list(value="B"))
}
