% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{tg.fraction}
\alias{tg.fraction}
\title{Cell Fraction Conversion}
\usage{
\method{tg}{fraction}(x, row, column, ...)
}
\arguments{
\item{x}{The fraction object to convert to a tg cell}

\item{row}{The AST row of that is generating this cell}

\item{column}{The AST column that is generating this cell}

\item{...}{additional specifiers for identifying this cell (see key)}
}
\value{
an S3 rendereable cell that is a hypothesis test
}
\description{
Construct a cell from a tg_fraction
}
\examples{
tg(tg_fraction(1, 2, 3), list(value="A"), list(value="B"))
}
