% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{tg.N}
\alias{tg.N}
\title{N value as cell}
\usage{
\method{tg}{N}(x, row, column, ...)
}
\arguments{
\item{x}{The tg_N object to convert into a rendereable cell}

\item{row}{The AST row of that is generating this cell}

\item{column}{The AST column that is generating this cell}

\item{...}{additional specifiers for identifying this cell (see key)}
}
\value{
an S3 rendereable cell label
}
\description{
Construct a cell from an N value.
}
\examples{
tg(tg_N("Joe"), list(value="A"), list(value="B"))
}
