% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{tg.na}
\alias{tg.na}
\title{Cell from NA}
\usage{
\method{tg}{na}(...)
}
\arguments{
\item{...}{arguments are ignored.}
}
\value{
an S3 rendereable cell
}
\description{
Construct a cell from NA
}
\examples{
tg(NA)
}

