% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{tg.numeric}
\alias{tg.numeric}
\title{Numeric Cell passed in}
\usage{
\method{tg}{numeric}(x, row, column, ...)
}
\arguments{
\item{x}{The numeric value to return as a renderable cell}

\item{row}{The AST row of that is generating this cell}

\item{column}{The AST column that is generating this cell}

\item{...}{additional specifiers for identifying this cell (see key)}
}
\value{
an S3 rendereable cell
}
\description{
Construct a cell from a numeric. This is essentially an identity function.
I.e., if a user has constructed a cell and passed it in do nothing.
}
\examples{
tg(23.0, list(value="A"), list(value="B"))
}
