% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{tg.quantile}
\alias{tg.quantile}
\title{Construct a cell from a tg_quantile}
\usage{
\method{tg}{quantile}(x, row, column, ...)
}
\arguments{
\item{x}{The quantile object to turn in to a rendereable cell}

\item{row}{The AST row of that is generating this cell}

\item{column}{The AST column that is generating this cell}

\item{...}{additional specifiers for identifying this cell (see key)}
}
\value{
an S3 rendereable cell that is a hypothesis test
}
\description{
Construct a cell from a tg_quantile
}
\examples{
tg(tg_quantile(rnorm(10), "\%.f"), list(value="A"), list(value="B"))
}
