% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{tg}
\alias{tg}
\title{Convert to Cell}
\usage{
tg(x, row, column, ...)
}
\arguments{
\item{x}{element to convert}

\item{row}{The AST row of that is generating this cell}

\item{column}{The AST column that is generating this cell}

\item{...}{additional specifiers for identifying this cell (see key)}
}
\value{
an S3 rendereable cell
}
\description{
Base S3 function to allow for constructing table cells. The main purpose
of this layer is to append additional traceability information, and convert
to internal cell type for later rendering decisions.
}
\examples{
tg(NA, list(value="A"), list(value="B"))
}
