% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{tg_fraction}
\alias{tg_fraction}
\title{Create a fraction that is convertible to a cell}
\usage{
tg_fraction(numerator, denominator)
}
\arguments{
\item{...}{the N values}
}
\value{
an S3 rendereable cell that is a hypothesis test
}
\description{
Create a fraction that is convertible to a cell
}
\examples{
tg_fraction(1, 2)
}

