% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{tg_N}
\alias{tg_N}
\title{N values creation}
\usage{
tg_N(...)
}
\arguments{
\item{...}{the N values}
}
\value{
an S3 rendereable cell that is a hypothesis test
}
\description{
Create a vector of N values that are convertable to a cell
}
\examples{
tg_N(1, 2, 3)
}
