% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{tg_quantile}
\alias{tg_quantile}
\title{Quantile creation}
\usage{
tg_quantile(x, ...)
}
\arguments{
\item{x}{the data passed to the {base}quantile function}

\item{...}{all arguments that are passed to the {base}quantile function}
}
\value{
an S3 rendereable cell that is a hypothesis test
}
\description{
Create a quantile that is convertible to a cell
}
\examples{
tg_fraction(1, 2)
}

