% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\docType{class}
\name{Token}
\alias{Token}
\title{A token in the formula grammar}
\format{\code{\link{R6Class}} object.}
\usage{
Token
}
\description{
A token in the formula grammar
}
\section{Fields}{

\describe{
\item{\code{id}}{The token identifier, E.g. "LPAREN"}

\item{\code{name}}{Information about the token, useful with IDENTIFIERs.}
}}

\section{Methods}{

\describe{
  \item{\code{new(id, name="")}}{Create a new token in the grammar.}
}
}

\keyword{data}
