% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{write_cell}
\alias{write_cell}
\title{Write a single cell}
\usage{
write_cell(table_builder, x, ...)
}
\arguments{
\item{table_builder}{The table builder to work on}

\item{x}{the cell to write}

\item{...}{additional attributes to pass for traceback}
}
\value{
a table builder with the given cell written in the current cursor position
}
\description{
Function to write a value to the current position in the table builder
}
\examples{
library(magrittr)
x <- Parser$new()$run(y ~ x)
new_table_builder(x$right, x$left) \%>\% write_cell(tg_N(23))

}
