% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-operators.R
\name{args_flatten}
\alias{args_flatten}
\title{Flatten variable arguments}
\usage{
args_flatten(...)
}
\arguments{
\item{...}{variable arguments}
}
\value{
a list of the arguments, with vectors and lists flattened
}
\description{
Take variable arguments, flatten vectors and lists, but do not flatten cells (which are lists)
e.g. args_flatten(NA, list(1,2,3), 4:6, c(7,8,9))
}
