% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\docType{class}
\name{ASTBranch}
\alias{ASTBranch}
\title{A left/right branch in an Abstract Syntrax Tree. This inherits from ASTNode, and
is intended to be a base class as well. Should never be instantiated directly
as once again the semantic information is contained in the class name.}
\format{
\code{\link{R6Class}} object.
}
\description{
A left/right branch in an Abstract Syntrax Tree. This inherits from ASTNode, and
is intended to be a base class as well. Should never be instantiated directly
as once again the semantic information is contained in the class name.

A left/right branch in an Abstract Syntrax Tree. This inherits from ASTNode, and
is intended to be a base class as well. Should never be instantiated directly
as once again the semantic information is contained in the class name.
}
\keyword{data}
\section{Super class}{
\code{\link[tangram:ASTNode]{tangram::ASTNode}} -> \code{ASTBranch}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{left}}{A pointer to the left node below this one}

\item{\code{right}}{A pointer to the right node below this one}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ASTBranch-distribute}{\code{ASTBranch$distribute()}}
\item \href{#method-ASTBranch-reduce}{\code{ASTBranch$reduce()}}
\item \href{#method-ASTBranch-clone}{\code{ASTBranch$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tangram" data-topic="ASTNode" data-id="set_format"><a href='../../tangram/html/ASTNode.html#method-ASTNode-set_format'><code>tangram::ASTNode$set_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tangram" data-topic="ASTNode" data-id="string"><a href='../../tangram/html/ASTNode.html#method-ASTNode-string'><code>tangram::ASTNode$string()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tangram" data-topic="ASTNode" data-id="terms"><a href='../../tangram/html/ASTNode.html#method-ASTNode-terms'><code>tangram::ASTNode$terms()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ASTBranch-distribute"></a>}}
\if{latex}{\out{\hypertarget{method-ASTBranch-distribute}{}}}
\subsection{Method \code{distribute()}}{
Call to distribute multiplication nodes, just recursively calls left and right node distribute functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ASTBranch$distribute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ASTBranch-reduce"></a>}}
\if{latex}{\out{\hypertarget{method-ASTBranch-reduce}{}}}
\subsection{Method \code{reduce()}}{
Attached data to nodes by processing data.frame appropriatly. Recursively calls left and right nodes to reduces on data.frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ASTBranch$reduce(df)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{(data.frame) Data frame to reduce over}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ASTBranch-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ASTBranch-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ASTBranch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
