% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\docType{class}
\name{ASTMultiply}
\alias{ASTMultiply}
\title{The multiplication of two terms, as an ASTNode.}
\format{
\code{\link{R6Class}} object.
}
\description{
The multiplication of two terms, as an ASTNode.

The multiplication of two terms, as an ASTNode.
}
\examples{
ASTMultiply$new(ASTVariable$new("x"), ASTVariable$new("y"))$string()

}
\keyword{data}
\section{Super classes}{
\code{\link[tangram:ASTNode]{tangram::ASTNode}} -> \code{\link[tangram:ASTBranch]{tangram::ASTBranch}} -> \code{ASTMultiply}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{left}}{The AST tree to the left.}

\item{\code{right}}{The AST tree to the right.}

\item{\code{type}}{The specified type of this node}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ASTMultiply-new}{\code{ASTMultiply$new()}}
\item \href{#method-ASTMultiply-distribute}{\code{ASTMultiply$distribute()}}
\item \href{#method-ASTMultiply-factors}{\code{ASTMultiply$factors()}}
\item \href{#method-ASTMultiply-string}{\code{ASTMultiply$string()}}
\item \href{#method-ASTMultiply-clone}{\code{ASTMultiply$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tangram" data-topic="ASTNode" data-id="set_format"><a href='../../tangram/html/ASTNode.html#method-ASTNode-set_format'><code>tangram::ASTNode$set_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tangram" data-topic="ASTNode" data-id="terms"><a href='../../tangram/html/ASTNode.html#method-ASTNode-terms'><code>tangram::ASTNode$terms()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tangram" data-topic="ASTBranch" data-id="reduce"><a href='../../tangram/html/ASTBranch.html#method-ASTBranch-reduce'><code>tangram::ASTBranch$reduce()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ASTMultiply-new"></a>}}
\if{latex}{\out{\hypertarget{method-ASTMultiply-new}{}}}
\subsection{Method \code{new()}}{
Construct a multiplication node
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ASTMultiply$new(left, right)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{left}}{(ASTNode) nodes to the left of the multiplication}

\item{\code{right}}{(ASTNode) nodes to the right of the multiplication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ASTMultiply-distribute"></a>}}
\if{latex}{\out{\hypertarget{method-ASTMultiply-distribute}{}}}
\subsection{Method \code{distribute()}}{
Rearrange nodes distribution multiplication across parenthesis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ASTMultiply$distribute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ASTMultiply-factors"></a>}}
\if{latex}{\out{\hypertarget{method-ASTMultiply-factors}{}}}
\subsection{Method \code{factors()}}{
return all terminal nodes on left and right
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ASTMultiply$factors()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ASTMultiply-string"></a>}}
\if{latex}{\out{\hypertarget{method-ASTMultiply-string}{}}}
\subsection{Method \code{string()}}{
Return a re-parseable string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ASTMultiply$string()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ASTMultiply-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ASTMultiply-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ASTMultiply$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
