% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\docType{class}
\name{ASTPlus}
\alias{ASTPlus}
\title{The addition of two terms, in an ASTNode.}
\format{
\code{\link{R6Class}} object.
}
\description{
The addition of two terms, in an ASTNode.

The addition of two terms, in an ASTNode.
}
\examples{
ASTPlus$new(ASTVariable$new("x"), ASTVariable$new("y"))$string()

}
\keyword{data}
\section{Super classes}{
\code{\link[tangram:ASTNode]{tangram::ASTNode}} -> \code{\link[tangram:ASTBranch]{tangram::ASTBranch}} -> \code{ASTPlus}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Just returns the R6 name 'ASTPlus'}

\item{\code{left}}{The node to the left of this node}

\item{\code{right}}{The node to the right of this node}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ASTPlus-new}{\code{ASTPlus$new()}}
\item \href{#method-ASTPlus-terms}{\code{ASTPlus$terms()}}
\item \href{#method-ASTPlus-string}{\code{ASTPlus$string()}}
\item \href{#method-ASTPlus-clone}{\code{ASTPlus$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tangram" data-topic="ASTNode" data-id="set_format"><a href='../../tangram/html/ASTNode.html#method-ASTNode-set_format'><code>tangram::ASTNode$set_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tangram" data-topic="ASTBranch" data-id="distribute"><a href='../../tangram/html/ASTBranch.html#method-ASTBranch-distribute'><code>tangram::ASTBranch$distribute()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tangram" data-topic="ASTBranch" data-id="reduce"><a href='../../tangram/html/ASTBranch.html#method-ASTBranch-reduce'><code>tangram::ASTBranch$reduce()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ASTPlus-new"></a>}}
\if{latex}{\out{\hypertarget{method-ASTPlus-new}{}}}
\subsection{Method \code{new()}}{
Construct a new node that represents addition
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ASTPlus$new(left, right)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{left}}{(ASTNode) Node on the left side of the addition}

\item{\code{right}}{(ASTNode) Node on the right side of the addition}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ASTPlus-terms"></a>}}
\if{latex}{\out{\hypertarget{method-ASTPlus-terms}{}}}
\subsection{Method \code{terms()}}{
Returns a vector of the left and right terms
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ASTPlus$terms()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ASTPlus-string"></a>}}
\if{latex}{\out{\hypertarget{method-ASTPlus-string}{}}}
\subsection{Method \code{string()}}{
A reparsable string representation of this node.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ASTPlus$string()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ASTPlus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ASTPlus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ASTPlus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
