% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\docType{class}
\name{ASTTableFormula}
\alias{ASTTableFormula}
\title{The root ASTNode of a formula.}
\format{
\code{\link{R6Class}} object.
}
\description{
The root ASTNode of a formula.

The root ASTNode of a formula.
}
\examples{
ASTTableFormula$new(ASTVariable$new("x"), ASTVariable$new("y"))$string()

}
\keyword{data}
\section{Super classes}{
\code{\link[tangram:ASTNode]{tangram::ASTNode}} -> \code{\link[tangram:ASTBranch]{tangram::ASTBranch}} -> \code{ASTTableFormula}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{left}}{The AST tree for the columns.}

\item{\code{right}}{The AST tree for the rows.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ASTTableFormula-new}{\code{ASTTableFormula$new()}}
\item \href{#method-ASTTableFormula-terms}{\code{ASTTableFormula$terms()}}
\item \href{#method-ASTTableFormula-string}{\code{ASTTableFormula$string()}}
\item \href{#method-ASTTableFormula-clone}{\code{ASTTableFormula$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tangram" data-topic="ASTNode" data-id="set_format"><a href='../../tangram/html/ASTNode.html#method-ASTNode-set_format'><code>tangram::ASTNode$set_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tangram" data-topic="ASTBranch" data-id="distribute"><a href='../../tangram/html/ASTBranch.html#method-ASTBranch-distribute'><code>tangram::ASTBranch$distribute()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tangram" data-topic="ASTBranch" data-id="reduce"><a href='../../tangram/html/ASTBranch.html#method-ASTBranch-reduce'><code>tangram::ASTBranch$reduce()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ASTTableFormula-new"></a>}}
\if{latex}{\out{\hypertarget{method-ASTTableFormula-new}{}}}
\subsection{Method \code{new()}}{
Create a new formula node
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ASTTableFormula$new(left, right)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{left}}{The left side of the "~" as an AST}

\item{\code{right}}{The right side of the "~" as an AST}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ASTTableFormula-terms"></a>}}
\if{latex}{\out{\hypertarget{method-ASTTableFormula-terms}{}}}
\subsection{Method \code{terms()}}{
Returns all terminal nodes from left and right
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ASTTableFormula$terms()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ASTTableFormula-string"></a>}}
\if{latex}{\out{\hypertarget{method-ASTTableFormula-string}{}}}
\subsection{Method \code{string()}}{
A re-parseable string representing the AST
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ASTTableFormula$string()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ASTTableFormula-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ASTTableFormula-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ASTTableFormula$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
