% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\docType{class}
\name{ASTTableFormula}
\alias{ASTTableFormula}
\title{The root ASTNode of a formula.}
\format{\code{\link{R6Class}} object.}
\usage{
ASTTableFormula
}
\description{
The root ASTNode of a formula.
}
\section{Fields}{

\describe{
\item{\code{left}}{The AST tree for the columns.}

\item{\code{right}}{The AST tree for the rows.}
}}

\section{Methods}{

\describe{
  \item{\code{new(left, right)}}{Create addition node of given left and right node.}
  \item{\code{terms()}}{Returns the a list of the left hand terms and right hand terms}
  \item{\code{distribute()}}{Applies the distributive property to the node, and returns the resulting node. This is the actual workhorse of the disributing multiplication across the tree.}
  \item{\code{string()}}{Returns the string representation of the formula}
  \item{\code{reduce(data)}}{Given a set of data, perform the logical reduction of the entire AST.}
}
}

\examples{
ASTTableFormula$new(ASTVariable$new("x"), ASTVariable$new("y"))$string()

}
\keyword{data}
