% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-cell.R
\name{cell.aov}
\alias{cell.aov}
\title{AOV model as cell}
\usage{
\method{cell}{aov}(x, pformat = "\%1.3f", ...)
}
\arguments{
\item{x}{The aov object to turn into a renderable cell}

\item{pformat}{numeric or character; A formatting directive to be applied to p-values}

\item{...}{additional specifiers for identifying this cell (see key)}
}
\value{
an S3 rendereable cell that is an F-statistic
}
\description{
Construct a cell from an analysis of variance model
}
\examples{
cell(aov(x~y,data.frame(x=rnorm(10), y=rnorm(10))))
}
