% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-cell.R
\name{cell_estimate}
\alias{cell_estimate}
\title{Create a cell_estimate object of the given estimate}
\usage{
cell_estimate(value, low, high, name = NULL, class = NULL, sep = ", ",
  ...)
}
\arguments{
\item{value}{The value of the estimate}

\item{low}{Specifies a lower interval for the estimate.}

\item{high}{Specifies an upper interval for the estimate.}

\item{name}{character; An optional name to apply to the value}

\item{class}{character; additional classs to apply}

\item{sep}{character; option separator character for the range}

\item{...}{optional extra information to attach}
}
\value{
A cell_estimate object.
}
\description{
A cell_estimate object contains a statistical estimate. It may
additionally contain an interval with a low and high of a
specified width.
}
\examples{
cell_estimate(1.0, 0.5, 1.5)
cell_estimate(1.0, 0.5, 1.5, name="one")
}
