% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-Cell.R
\name{cell_estimate}
\alias{cell_estimate}
\title{Create an cell_estimate (S3) object of the given estimate}
\usage{
cell_estimate(value, low = NA, high = NA, conf.level = 0.95, src = NA)
}
\arguments{
\item{value}{The value of the estimate}

\item{low}{An optional field that specifies a lower interval for the estimate.}

\item{high}{An optional field that specifies an upper interval for the estimate.}

\item{conf.level}{An optional field for storing the width of the interval.}

\item{src}{An optional field for traceability of the source of the field}
}
\value{
A cell_estimate object.
}
\description{
A cell_estimate object contains a statistical estimate. It may
additionally contain an interval with a low and high of a
specified width.
}
\examples{
cell_estimate(1.0, 0.5, 1.5)
}
