% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-cell.R
\name{cell_fraction}
\alias{cell_fraction}
\title{Create an cell_fraction (S3) object of the given statistic}
\usage{
cell_fraction(numerator, denominator, format = 3, class = NULL, ...)
}
\arguments{
\item{numerator}{numeric; The value of the numerator}

\item{denominator}{numeric; The value of the denominator}

\item{format}{numeric or character; a string formatting directive}

\item{class}{character; An optional field for additional S3 classes (e.g. could be used in html rendering for CSS)}

\item{...}{optional extra information to attach}
}
\value{
A cell_fraction object.
}
\description{
A cell_fraction object contains a statistical result of a fraction/percentage.
}
\examples{
cell_fraction(1, 4, 0.25, 25)
}
