% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-cell.R
\name{cell_fstat}
\alias{cell_fstat}
\title{Create an cell_fstat (S3) object of the given statistic}
\usage{
cell_fstat(f, df1, df2, p, class = NULL, ...)
}
\arguments{
\item{f}{The value of the f-statistic}

\item{df1}{1st dimension degrees of freedom}

\item{df2}{2nd dimension degrees of freedom}

\item{p}{The p-value of the resulting test}

\item{class}{character; An optional field for additional S3 classes (e.g. could be used in html rendering for CSS)}

\item{...}{optional extra information to attach}
}
\value{
A cell_fstat object.
}
\description{
A cell_fstat object contains a statistical result of an F-test.
}
\examples{
cell_fstat(4.0, 10, 20, 0.004039541, reference=1,)
}
