% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-Cell.R
\name{cell_header}
\alias{cell_header}
\title{Create a cell_header (S3) object of the given text.}
\usage{
cell_header(text, units = NA, src = NA)
}
\arguments{
\item{text}{The text of the label}

\item{units}{An optional field that contains units}

\item{src}{An optional field for traceability of the source of the field}
}
\value{
A cell_header object
}
\description{
A cell_header object is a sub class of a cell_label representing
that this label is a header and upon rendering may invoke different
handling.
}
\examples{
cell_header("Bold THIS")
}
