% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-Cell.R
\name{cell_label}
\alias{cell_label}
\title{Create an cell_label (S3) object of the given text.}
\usage{
cell_label(text, units = NA, src = NA)
}
\arguments{
\item{text}{The text of the label}

\item{units}{An optional field that contains units}

\item{src}{An optional field for traceability of the source of the field}
}
\value{
A cell_table object
}
\description{
A cell_label object represents a label cell inside a table. It can
also contain units.
}
\examples{
cell_label("Compaction Method")
}
