% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-cell.R
\name{cell_n}
\alias{cell_n}
\title{Create an cell_n (S3) object of the given statistic}
\usage{
cell_n(n, class = NULL, hdr = FALSE, possible = NULL, ...)
}
\arguments{
\item{n}{The numerical value}

\item{class}{character; An optional field for additional S3 classes (e.g. could be used in html rendering for CSS)}

\item{hdr}{logical; Construct an n value for a header (defaults to FALSE)}

\item{possible}{numerical; The total N that was possible}

\item{...}{optional extra information to attach}
}
\value{
A cell_n object.
}
\description{
A cell_n object contains an n value. Essentially, this is just a helper that appends the cell_n class to the given
object and makes sure it's a cell S3 object as well.
}
\examples{
cell_n(20)
}
