% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-cell.R
\name{cell_n}
\alias{cell_n}
\title{Create an cell_n (S3) object of the given statistic}
\usage{
cell_n(n, class = NULL, ...)
}
\arguments{
\item{n}{The value of the X^2 statistic}

\item{class}{character; An optional field for additional S3 classes (e.g. could be used in html rendering for CSS)}

\item{...}{optional extra information to attach}
}
\value{
A cell_n object.
}
\description{
A cell_n object contains a statistical result of an X^2-test.
}
\examples{
cell_n(20)
}
