% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-cell.R
\name{cell_spearman}
\alias{cell_spearman}
\title{Create an cell_spearman (S3) object of the given statistic}
\usage{
cell_spearman(S, rho, p, class = NULL, ...)
}
\arguments{
\item{S}{The value of the spearman statistic}

\item{rho}{The rho value of the test}

\item{p}{p-value of resulting test}

\item{class}{character; An optional field for additional S3 classes (e.g. could be used in html rendering for CSS)}

\item{...}{optional extra information to attach}
}
\value{
A cell_spearman object.
}
\description{
A cell_spearman object contains a statistical result of an spearman-test.
}
\examples{
cell_spearman(20, 0.2, 0.05)
}
