% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-Cell.R
\name{cell_studentt}
\alias{cell_studentt}
\title{Create an cell_studentt (S3) object of the given statistic}
\usage{
cell_studentt(t, df, p, reference = NA, src = NA)
}
\arguments{
\item{t}{The value of the X^2 statistic}

\item{df}{degrees of freedom}

\item{p}{p-value of resulting test}

\item{reference}{A possible reference number for use in a table key}

\item{src}{An optional field for traceability of the source of the field}
}
\value{
A cell_studentt object.
}
\description{
A cell_studentt object contains a statistical result of an t-test.
}
\examples{
cell_studentt(2.0, 20, 0.02963277)
}
