% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-cell.R
\name{cell_subheader}
\alias{cell_subheader}
\title{Create a cell_subheader object of the given text.}
\usage{
cell_subheader(text, units = NULL, class = NULL, ...)
}
\arguments{
\item{text}{character; The text of the label. May include a subset of LaTeX greek or math.}

\item{units}{character; An optional field that contains units}

\item{class}{character; An optional field for additional S3 classes (e.g. could be used in html rendering for CSS)}

\item{...}{optional extra information to attach}
}
\value{
A cell_subheader object.
}
\description{
A cell_subheader object represents a label cell inside a table. It can
also contain units.
}
\examples{
cell_subheader("Concentration")
cell_subheader("Concentration", "mg/dl")
cell_subheader("Concentration", "mg/dl", src="A")
}
