% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cell_transform}
\alias{cell_transform}
\title{Create a function to transform all cells of a table}
\usage{
cell_transform(FUN, ...)
}
\arguments{
\item{FUN}{function to apply, must return the modified cell}

\item{...}{additional arguments to pass into function}
}
\value{
a table modification function
}
\description{
Given a function that operates on a table cell and returns
the modified cell, return a function that given a table
applies that function to all cells and returns the modified
table.
}
