% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{cursor_up}
\alias{cursor_up}
\title{Move cursor up}
\usage{
cursor_up(table_builder, n = 1)
}
\arguments{
\item{table_builder}{The table builder to work on}

\item{n}{units to move cursor up}
}
\value{
a table builder with the cursor up n positions
}
\description{
Move table builder cursor up specified value (default 1)
}
\examples{
library(magrittr)
x <- Parser$new()$run(y ~ x)
new_table_builder(x$right, x$left) \%>\% cursor_pos(3,3) \%>\% cursor_up(2)

}
