% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{derive_label}
\alias{derive_label}
\title{Derive label of AST node.}
\usage{
derive_label(node)
}
\arguments{
\item{node}{Abstract syntax tree node.}
}
\value{
A string with a label for the node
}
\description{
Determine the label of a given AST node.
NOTE: Should have data attached via reduce before calling.
}
