% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{form}
\alias{form}
\title{Attach format attribute to object}
\usage{
form(object, value)
}
\arguments{
\item{object}{the object to attach format information}

\item{value}{the formatting to be applied (usually comes from AST node)}
}
\value{
an S3 object with format attribute set
}
\description{
Attach formatting information to an object in the attr "format"
}
\examples{
form(2, "\%.2f")
}
