% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{format_guess}
\alias{format_guess}
\title{Guess the best format for a set of data}
\usage{
format_guess(x)
}
\arguments{
\item{x}{the data passed to the quantile {stats} function}
}
\value{
the significant digits to preserve
}
\description{
Given a vector of data, default to 3 significant digits or all if maximum is greater
than zero
}
\examples{
format_guess(rnorm(100))
}
