% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-hmisc.R
\docType{data}
\name{hmisc_style}
\alias{hmisc_style}
\title{Style Bundle for Hmisc defaults.}
\format{An object of class \code{list} of length 5.}
\usage{
hmisc_style
}
\description{
List of lists, should contain a "Type" entry with a function to determine type of vector passed in.
Next entries are keyed off returned types from function, and represent the type of a row.
The returned list should contain the same list of types, and represents the type of a column. Thus it now returns
a function to process the intersection of those two types.
}
\keyword{data}
