% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index.cell_fstat}
\alias{index.cell_fstat}
\title{Generate an an index from a cell_fstat object}
\usage{
\method{index}{cell_fstat}(object, caption, ...)
}
\arguments{
\item{object}{The cell_fstat for indexing}

\item{caption}{an additional specifier for the object key}

\item{...}{additional arguments to renderer. Unused}
}
\value{
A matrix of strings containing key, source and value
}
\description{
Given a cell_fstat class create an index representation. If no source
is specified no index will be generated.
}
