% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index.default}
\alias{index.default}
\title{Generate an index from a cell object}
\usage{
\method{index}{default}(object, caption, ...)
}
\arguments{
\item{object}{The cell for indexing}

\item{caption}{an additional specifier for the object key}

\item{...}{additional arguments to renderer. Unused}
}
\value{
A matrix of strings containing key, source and value
}
\description{
Given a cell class create an index representation. If no source
is specified no index will be generated.
}
