% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-index.R
\name{index}
\alias{index}
\title{Generate an index from a tangram or cell object}
\usage{
index(object, ...)
}
\arguments{
\item{object}{The cell header to render to HTML5}

\item{...}{additional arguments to renderer. Unused}
}
\value{
A matrix or list of strings containing key, source and value
}
\description{
Given a tangram object create an index representation.
}
