% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-index.R
\name{index.tangram}
\alias{index.tangram}
\title{Generate an an index from a tangram object}
\usage{
\method{index}{tangram}(object, id = "tangram", key.len = 4, ...)
}
\arguments{
\item{object}{The tangram for indexing}

\item{id}{an additional specifier for the object key}

\item{key.len}{numeric; length of keys generated (affects collision probability)}

\item{...}{additional arguments to renderer. Unused}
}
\value{
A matrix of strings containing key, source and value
}
\description{
Given a tangram class create an index representation.
}
