% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-latex.R
\name{latex}
\alias{latex}
\alias{latex.default}
\alias{latex.cell}
\alias{latex.cell_label}
\alias{latex.cell_n}
\alias{latex.cell_header}
\alias{latex.cell_subheader}
\alias{latex.tangram}
\title{Render to LaTeX methods for tangram cell objects}
\usage{
latex(object, ...)

\method{latex}{default}(object, ...)

\method{latex}{cell}(object, ...)

\method{latex}{cell_label}(object, ...)

\method{latex}{cell_n}(object, ...)

\method{latex}{cell_header}(object, ...)

\method{latex}{cell_subheader}(object, ...)

\method{latex}{tangram}(object, caption = "Table", footnote = NULL,
  fragment = FALSE, filename = NULL, append = FALSE, na.blank = TRUE,
  cgroup.just = NULL, ...)
}
\arguments{
\item{object}{object; the item to render to latex}

\item{...}{additional arguments}
}
\value{
the LaTeX rendering
}
\description{
Each of these methods will render the cell object as a LaTeX fragment
}
\examples{
latex(cell_label("123"))
latex(cell_iqr(rnorm(20)))
latex(cell_estimate(2.1,0.8, 3.3))
latex(cell_fraction(45, 137))
latex(table_builder()   \%>\%
        row_header("row") \%>\%
        col_header(1,2,3) \%>\%
        add_col("A","B","C"))
latex(tangram(drug~bili, pbc))
}
