% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{line_feed}
\alias{line_feed}
\title{Move cursor to next line
Move table builder cursor to the next line (does not alter column)}
\usage{
line_feed(table_builder, n = 1)
}
\arguments{
\item{table_builder}{The table builder to work on}

\item{n}{optional number of line_feeds to perform}
}
\value{
a table builder with the cursor at the first column
}
\description{
Move cursor to next line
Move table builder cursor to the next line (does not alter column)
}
\examples{
library(magrittr)
x <- Parser$new()$run(y ~ x)
new_table_builder(x$right, x$left) \%>\% cursor_pos(3,3) \%>\% line_feed()

}
