% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{new_col}
\alias{new_col}
\title{Open a new column in 1st row}
\usage{
new_col(table_builder)
}
\arguments{
\item{table_builder}{The table builder to work on}
}
\value{
a table builder with the cursor at the first column
}
\description{
Advance table builder cursor to the furthest right column on the top row and open a new column
}
\examples{
library(magrittr)
x <- Parser$new()$run(y ~ x)
new_table_builder(x$right, x$left) \%>\% new_col()

}
