% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{new_row}
\alias{new_row}
\title{Open a new row}
\usage{
new_row(table_builder)
}
\arguments{
\item{table_builder}{The table builder to work on}
}
\value{
a table builder with the cursor at the first column
}
\description{
Move table builder cursor to the bottom of all defined rows opening a new one
in the first column
}
\examples{
library(magrittr)
x <- Parser$new()$run(y ~ x)
new_table_builder(x$right, x$left) \%>\% new_row()

}
