% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-summary.R
\name{print.cell}
\alias{print.cell}
\title{Print a text summary of a given table}
\usage{
\method{print}{cell}(x, ...)
}
\arguments{
\item{x}{The tangram to render to text}

\item{...}{additional arguments, unused at present}
}
\value{
A text string rendering of the given table
}
\description{
Print a text summary of a given table
}
