% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-rmd.R
\name{rmd}
\alias{rmd}
\title{Generate an Rmd table entry from a cell object}
\usage{
rmd(object, ...)
}
\arguments{
\item{object}{The cell_fstat for indexing}

\item{...}{additional arguments to renderer. Unused}
}
\value{
A string representation of the table
}
\description{
Given a cell object generate the corresponding piece of an Rmd table
}
