% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-cell.R
\name{rows}
\alias{rows}
\alias{cols}
\alias{rows.list}
\alias{cols.list}
\title{S3 object to return number of rows/cols in object}
\usage{
rows(x)

cols(x)

\method{rows}{list}(x)

\method{cols}{list}(x)
}
\arguments{
\item{x}{object; object to determine requested count}
}
\description{
Number of rows/cols in provided object
}
