% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-rtf.R
\name{rtf.cell_fstat}
\alias{rtf.cell_fstat}
\title{Convert an abstract cell_fstat object into an RTF string}
\usage{
\method{rtf}{cell_fstat}(object, id, ...)
}
\arguments{
\item{object}{The cell fstat to render to RTF}

\item{id}{A unique identifier for traceability}

\item{...}{additional arguments to renderer. Unused}
}
\value{
A text string rendering of the given fstat as a <td> with several <span>'s.
}
\description{
Given a cell_fstat class create an RTF representation.
}
