% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-rtf.R
\name{rtf.cell_label}
\alias{rtf.cell_label}
\title{Given a cell_label class create an RTF representation.}
\usage{
\method{rtf}{cell_label}(object, id, ..., point = 18)
}
\arguments{
\item{object}{The cell label to render to RTF}

\item{id}{A unique identifier for traceability}

\item{...}{additional arguments to renderer. Unused}

\item{point}{size of main font for cell label}
}
\value{
An RTF text string rendering of the given label.
}
\description{
Given a cell_label class create an RTF representation.
}
