% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-rtf.R
\name{rtf.cell_subheader}
\alias{rtf.cell_subheader}
\title{Convert an abstract cell_subheader object into an RTF string}
\usage{
\method{rtf}{cell_subheader}(object, id, ..., point = 9)
}
\arguments{
\item{object}{The cell header to render to RTF}

\item{id}{A unique identifier for traceability}

\item{...}{additional arguments to renderer. Unused}

\item{point}{numeric; The font point size to use in display}
}
\value{
An RTF string rendering of the given header
}
\description{
Given a cell_subheader class create an RTF representation.
}
