% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf.R
\name{rtf.cell_table}
\alias{rtf.cell_table}
\title{Convert an abstract table object into an RTF string or file}
\usage{
\method{rtf}{cell_table}(object, caption = NA, fragment = FALSE, id = NA,
  widths = NA, footnote = NA, filename = NA, append = FALSE, ...)
}
\arguments{
\item{object}{The cell table to render to RTF}

\item{caption}{A string caption for the table}

\item{fragment}{A boolean flag that determines whether a fragment or a complete RTF document is generatedf}

\item{id}{A unique identifier for the table (strongly recommended).}

\item{widths}{RTF requires specified left margin and column widths, this allows user control over these (inches)}

\item{footnote}{Any footnotes to include under the table.}

\item{filename}{A filename to write resulting rtf file to}

\item{append}{A boolean for whether or not to append to given filename}

\item{...}{additional arguments to renderer. Unused}
}
\value{
A text string rendering of the given table
}
\description{
Given a cell_table class, a series of conversion creates an rtf
representation of the table.
}
