% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf.R
\name{rtf}
\alias{rtf}
\title{S3 rtf Method function for use on abstract table class}
\usage{
rtf(object, ...)
}
\arguments{
\item{object}{The cell to render to RTF}

\item{...}{additional arguments to renderer.}
}
\description{
S3 rtf Method function for use on abstract table class
}
