% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-rtf.R
\name{rtf}
\alias{rtf}
\title{S3 rtf Method function for use on abstract table class}
\usage{
rtf(object, id, ...)
}
\arguments{
\item{object}{The cell to render to RTF}

\item{id}{A unique identifier for the table (strongly recommended). If not provided, caption will be used.}

\item{...}{additional arguments to renderer. Unused at present.}
}
\value{
A text string rendering of the given table
}
\description{
S3 rtf Method function for use on abstract table class
}
