% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmisc.R
\name{summarize_chisq}
\alias{summarize_chisq}
\title{Create a summarization for a categorical row versus a categorical column}
\usage{
summarize_chisq(table, row, column)
}
\arguments{
\item{table}{The table object to modify}

\item{row}{The row variable object to use (categorical)}

\item{column}{The column variable to use (categorical)}
}
\value{
The modified table object
}
\description{
Given a row and column object from the parser apply a chi^2 test and output
the results
}
