% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-hmisc.R
\name{summarize_chisq}
\alias{summarize_chisq}
\title{Create a summarization for a categorical row versus a categorical column}
\usage{
summarize_chisq(table, row, column, pformat = NULL)
}
\arguments{
\item{table}{The table object to modify}

\item{row}{The row variable object to use (categorical)}

\item{column}{The column variable to use (categorical)}

\item{pformat}{numeric or character; A formatting directive to be applied to p-values}
}
\value{
The modified table object
}
\description{
Given a row and column object from the parser apply a chi^2 test and output
the results
}
