% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmisc.R
\name{summarize_chisq_single}
\alias{summarize_chisq_single}
\title{Create a summarization for a binomial row versus a categorical column}
\usage{
summarize_chisq_single(table, row, column)
}
\arguments{
\item{table}{The table object to modify}

\item{row}{The row variable object to use (binomial)}

\item{column}{The column variable to use (categorical)}
}
\value{
The modified table object
}
\description{
Given a row and column object from the parser apply a chi^2 test and output
the results (1) X (N,#col categories ,statistic) format.
}
