% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmisc.R
\name{summarize_kruskal_horz}
\alias{summarize_kruskal_horz}
\title{Create a summarization for a categorical set of column versus a numerical row}
\usage{
summarize_kruskal_horz(table, row, column)
}
\arguments{
\item{table}{The table object to modify}

\item{row}{The row variable object to use (numerical)}

\item{column}{The column variable to use (categorical)}
}
\value{
The modified table object
}
\description{
Given a row and column object from the parser apply a Kruskal test and output
the results horizontally. 1 X (n + no. categories + test statistic)
}
