% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-hmisc.R
\name{summarize_spearman}
\alias{summarize_spearman}
\title{Create a summarization for a numerical row versus a numerical column}
\usage{
summarize_spearman(table, row, column, pformat = NULL)
}
\arguments{
\item{table}{The table object to modify}

\item{row}{The row variable object to use (numerical)}

\item{column}{The column variable to use (numerical)}

\item{pformat}{numeric or character; A formatting directive to be applied to p-values}
}
\value{
The modified table object
}
\description{
Given a row and column object from the parser apply a Spearman test and output
the results in a 1X3 format.
}
