% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{tg.aov}
\alias{tg.aov}
\title{AOV model as cell}
\usage{
\method{tg}{aov}(x, row, column, ...)
}
\arguments{
\item{x}{The aov object to turn into a renderable cell}

\item{row}{The AST row of that is generating this cell}

\item{column}{The AST column that is generating this cell}

\item{...}{additional specifiers for identifying this cell (see key)}
}
\value{
an S3 rendereable cell that is an F-statistic
}
\description{
Construct a cell from an analysis of variance model
}
\examples{
tg(aov(rnorm(10) ~ rnorm(10)), list(value="A"), list(value="B"))
}
