% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{tg.htest}
\alias{tg.htest}
\title{Construct hypothesis test form cell}
\usage{
\method{tg}{htest}(x, row, column, ...)
}
\arguments{
\item{x}{The htest object to convert to a rendereable cell}

\item{row}{The AST row of that is generating this cell}

\item{column}{The AST column that is generating this cell}

\item{...}{additional specifiers for identifying this cell (see key)}
}
\value{
an S3 rendereable cell that is a hypothesis test
}
\description{
Construct a cell from a hypothesis test
}
\examples{
tg(t.test(rnorm(10),rnorm(10)), list(value="A"), list(value="B"))
}
