% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{tg_quantile}
\alias{tg_quantile}
\title{Quantile creation}
\usage{
tg_quantile(x, format = NA, ...)
}
\arguments{
\item{x}{the data passed to the quantile {stats} function}

\item{format}{the formatting to be applied (usually comes from AST node)}

\item{...}{all arguments that are passed to the quantile {stats} function}
}
\value{
an S3 rendereable cell that is a hypothesis test
}
\description{
Create a quantile that is convertible to a cell
}
\examples{
tg_quantile(rnorm(100), "\%.2f")
}
